requirements.txt
asgiref==3.7.2
astroid==3.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
dill==0.3.8
Django==5.0.4
django-cors-headers==4.3.1
django-ipware==7.0.1
django-rest-authemail==2.1.7
djangorestframework==3.14.0
idna==3.7
isort==5.13.2
lxml==5.2.1
mccabe==0.7.0
platformdirs==4.2.0
pylint==3.1.0
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
pyparsing==2.4.7
python-ipware==3.0.0
pytz==2024.1
requests==2.31.0
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.12.4
typing_extensions==4.11.0
unittest-xml-reporting==3.2.0
urllib3==2.2.1
xmlrunner==1.7.7
spacy
pypdf
en-core-web-md==3.7.1
textacy